move new comobox to components folder
